[Refactor] - AppData : Properly checks if the server returned error response for exhibitions call and handles it with appropriate message
